third_party/blink/web_tests/fast/files/file-reader-fffd-expected.txt

Throw various bad bytes at file reader.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

array = new Uint8Array([65, 245, 246, 247, 248, 249, 250, 251, 252, 253, 254, 255, 66]);
blob = new Blob([array]);
reader = new FileReader();
reader.readAsText(blob);
PASS fileString is 'A�����������B'
PASS successfullyParsed is true

TEST COMPLETE
